Bugfix Mac Os compilation issues
In other case build fails with internal compiler error

+++ imageproc/Binarize.cpp
@@ -103,18 +103,18 @@ BinaryImage binarizeSauvola(QImage const& src, QSize const window_size)
 			assert(area > 0); // because window_size > 0 and w > 0 and h > 0
 			
 			QRect const rect(left, top, right - left, bottom - top);
-			double const window_sum = integral_image.sum(rect);
-			double const window_sqsum = integral_sqimage.sum(rect);
+			long double const window_sum = integral_image.sum(rect);
+			long double const window_sqsum = integral_sqimage.sum(rect);
 			
-			double const r_area = 1.0 / area;
-			double const mean = window_sum * r_area;
-			double const sqmean = window_sqsum * r_area;
+			long double const r_area = 1.0 / area;
+			long double const mean = window_sum * r_area;
+			long double const sqmean = window_sqsum * r_area;
 			
-			double const variance = sqmean - mean * mean;
-			double const deviation = sqrt(fabs(variance));
+			long double const variance = sqmean - mean * mean;
+			long double const deviation = sqrt(fabs(variance));
 			
-			double const k = 0.34;
-			double const threshold = mean * (1.0 + k * (deviation / 128.0 - 1.0));
+			long double const k = 0.34;
+			long double const threshold = mean * (1.0 + k * (deviation / 128.0 - 1.0));
 			
 			uint32_t const msb = uint32_t(1) << 31;
 			uint32_t const mask = msb >> (x & 31);
@@ -177,7 +177,7 @@ BinaryImage binarizeWolf(
 	std::vector<float> means(w * h, 0);
 	std::vector<float> deviations(w * h, 0);
 	
-	double max_deviation = 0;
+	long double max_deviation = 0;
 	
 	for (int y = 0; y < h; ++y) {
 		int const top = std::max(0, y - window_lower_half);
@@ -190,15 +190,15 @@ BinaryImage binarizeWolf(
 			assert(area > 0); // because window_size > 0 and w > 0 and h > 0
 			
 			QRect const rect(left, top, right - left, bottom - top);
-			double const window_sum = integral_image.sum(rect);
-			double const window_sqsum = integral_sqimage.sum(rect);
+			long double const window_sum = integral_image.sum(rect);
+			long double const window_sqsum = integral_sqimage.sum(rect);
 			
-			double const r_area = 1.0 / area;
-			double const mean = window_sum * r_area;
-			double const sqmean = window_sqsum * r_area;
+			long double const r_area = 1.0 / area;
+			long double const mean = window_sum * r_area;
+			long double const sqmean = window_sqsum * r_area;
 			
-			double const variance = sqmean - mean * mean;
-			double const deviation = sqrt(fabs(variance));
+			long double const variance = sqmean - mean * mean;
+			long double const deviation = sqrt(fabs(variance));
 			max_deviation = std::max(max_deviation, deviation);
 			means[w * y + x] = mean;
 			deviations[w * y + x] = deviation;
@@ -216,9 +216,9 @@ BinaryImage binarizeWolf(
 		for (int x = 0; x < w; ++x) {
 			float const mean = means[y * w + x];
 			float const deviation = deviations[y * w + x];
-			double const k = 0.3;
-			double const a = 1.0 - deviation / max_deviation;
-			double const threshold = mean - k * a * (mean - min_gray_level);
+			long double const k = 0.3;
+			long double const a = 1.0 - deviation / max_deviation;
+			long double const threshold = mean - k * a * (mean - min_gray_level);
 			
 			uint32_t const msb = uint32_t(1) << 31;
 			uint32_t const mask = msb >> (x & 31);
